FIX: using ->subtract breaks Translatable
Using $allList->subtract($gridField->getList()) breaks Translatable->augmentSQL because
it finds the Locale filter in the subquery and assumes it does not need to augment the
actual search query.  When you have a very large table with many records not in the
locale that you are searching, this can affect the search time by several orders of
magnitude.