MINOR: better debug information for mismatch in variables in Config class
The current mismatch in errors is clear about what is wrong, but provides no information on where the error was made, making it impossible to fix the error without adding code similar to the code added by this proposed change.  The proposed change basically shows the two values, which will make it easy for the developer to find the mismatched values and fix them.