MINOR: GridFieldDetailForm::doSave() returns unescaped HTML in success message
Due to the new default casting of Text, the HTML is unescaped. Citing that the anchor tag in the success message is of little utility to the user, this change removes the HTML and leaves a plain success message. Note that the {link} key in the translation no longer has any semantic meaning. Consider updating the lang files.