BUG #2553 Fixed how HTMLEditorField calls ResizedImage function to address capitalization issue
Added file check to HTMLEditorField saveInto function so that it would first check for an existing resized image, and compare its size to the the image tag before passing the File off to the Image Object to be resized. This will prevent SilverStripe from regenerating the img tag when the correct resized image already exists.