ENHANCEMENT: automatic link child class (has_one parent) in addNewForm - replaces 2637
Presets and locks-in the parent HasOne relationship value for a child record (e.g. if we add a city to a country then the user can see that the country is preset and the user can not change the country (as it will be overridden to the parent country anyway!).  

This has a number of advantage over the current situation that are discussed in the previous pull request. 


for a full description and discussion, see https://github.com/silverstripe/silverstripe-framework/pull/2637