Update Requirements.php
When including a Javascript or CSS file in a page, if path is incorrect, I suggest to send an error message in order to inform developers about the path. Currently, if some file fails to be included, there's no way to know what happened.

I am not sure if the message is the most suitable. Please, review it.

Thanks