allow themed renderWith even if theme is disabled
```php
$this->renderWith(foo)
```
called while the user is editing the CMS (e.g. when creating a newsletter) often leads to unexpected results: the theme of the site is not being used and therefore no themed templates can be found because the CMS turns of the site's theme (for good reasons).  

This proposed addition to the renderwith method allows devs to render in theme style even if called from within the CMS (of other situations where the theme is disabled).