BUG AssetUploadField hides "generic" file upload messages.
If an upload fails, the error is not shown correctly, it gets hidden
off the table because of absolute positioning. This currently worked
because it relied upon known errors not showing a "size" element next
to it.

"ui-state-warning-text" and "ui-state-success-text" don't need this
CSS, so it's probably not necessary anyway as it shows the error text
correctly without it.