FIX: Class names referenced in Config should be case-insensitive
The config system maps top-level key name to a PHP class. PHP classes are case-insensitive,
but without this patch, the Config system was case-sensitive on this variable. I think this
should be considered a bug.

The following patch makes the "class" key case-insensitive, but leave the "name" key, as
case-sensitive, mirroring the semantics of PHP classes and static variables, respectively.

Arguably, the entire config system could be case-insensitive, but I haven't done that in
this patch.