Allow setting cache lifetime for SS_Configuration and i18n caches
This currently hardcodes the lifetimes to null. Doing so will make the
cache last indefinitely. However, this means we can't override the
lifetime from custom code, like this:

	SS_Cache::set_cache_lifetime('any', 3600);

This change sets the cache lifetime configuration from
framework/_config.php so there's the potential to override it with
a higher priority from custom code.