Removing Makefile referring to old commands, no longer seems useful.
This Makefile doesn't appear to be that useful anymore, especially
considering it's running tests via `dev/tests/all`, an old way of
running the tests. It also refers to "windmill", which I'm pretty sure
has been removed, considering it refers to the "cms" module as well.

Continuous integration should simply just use the `phpunit` command
which is able to output better results for TeamCity, etc.

The only target here which is actually useful is the jasmine one, but
I think that could just be started manually, instead of using this
Makefile.