Fix TreeDropdownField label on deselect
When deselecting the currently selected item in a TreeDropdownField component, the select list label displays '()'. 
This is confusing user feedback as it does not convey the reset action that was intended by deselecting the item.
If the title is set to an empty string, the () is still displayed, so my workaround is to set the title to a single space.