Recipe on removing WYSISYG buttons
Currently, it is not possible to run multiple HtmlEditorConfigs easily on one site. Sometimes you may want to remove some buttons that aren't of use to the content editor for the task they are to performs. Herein lies an example of how to achieve removing buttons. (spent 5+ hours attempting to use different CustomHtmlEditorConfig modules from git etc), finally I got something to work, the setTimeout was the trick ;p