DOCS Mention the option of Environment files for WAMP users
Once you get a few sites under your belt most devs would likely turn to using environment files to names DB credentials. This introduces this a bit earlier to get them used to the simple idea of an environment file.

Eventually I'd like to rewrite the installer to generate this over storing in the _config.php.