FIX: Corrected padding on non-sortable columns.
span.non-sortable in the GridField header relies on there being other, sortable columns in the list to get the padding
right. If there are no such columns (e.g. in a simple header), then the header is shown without padding, which is ugly.

This small change fixes that. Tested on Chrome and Firefox. 1px smaller than a sortable header on Firefox but still looks fine.