NEW: Add support for ‘%%remove%%’ market in Yaml Config.
Sometimes, it’s not enough to set a config option to null. For example,
you may configure a list of items as a constructor property setter and
it won’t expect empty items in the string.

This change introduces a special string that you can use to mark config
options that you would like to have removed with unset().

Note that this means we can’t set a configuration option to the string
“%%remove%%”; hence the reason for using a slightly convoluted string.