BUGFIX: If ->loadDataFrom is given an array, allow multiples for namespaced keys
When `Form->loadDataFrom()` is passed an array for `$data` (e.g. what selecting
multiple values for a `CheckboxsetField` would do), this allows those values to
be correctly set by `loadDataFrom`.

I can probably provide a reproducible test case if required, for me it
manifested in a `GridField` inside a `ModelAdmin` `GridField` on a
`CheckboxsetField` which has a field name something like `Schedule[Daily][Days]`,
which means e.g. the first value would be `Schedule[Daily][Days][1] = 1`.