Don't run forceSSL() on CLI
Was performing an inefficient check for this previously.
REQUEST_URI is set through cli-script.php, hence it's not
an accurate way to determine CLI operation.

The effect of this bug is that command line tasks will fail to
run since they'll return a HTTP redirect which a CLI client can't interpret.