Adding a transparent background to upload preview
Adding a background to upload preview

I could not see a preview of a white image with background transparency in the preview area. Adding a backgorund to this area helps make it more visible.

Applying background image to img

Applying background image to img to avoid layout change

Applying background image to img

Applying background image to img to avoid layout change