Add special flag to CMS preview requests
This affects the iframe used in the CMS for page previews. With this change, we can have the page behave differently based on whether it's a preview or not. The page itself would have to check for the parameter, nothing is done automatically.

Our specific use case is to prevent redirecting to another page, as it would normally. However this may also be useful in e.g. plastering a PREVIEW message, or having other preview-specific content.