NEW Set action param on internal links
When inserting a link to an internal page you can now choose an action from that page's $allowed_actions. This works in a very similar fashion to anchor selection. You can include an ID or other params here too - basically anything that would come between the page URL and the '#' fragment identifier.