Fix the most egregious error for an NZ based proj.
The Bay of Plenty Rugby team is the Steamers, and play in the NPC (National Provincial Championship), hailing from the Te Ika a Maui (North Island).
The Canterbury Crusaders (note the alliteration) hail from Te Waipounamu (South Island), and are an entry in the Super Rugby championship, which is an international affair (also known as the Super12, Super15 & most recently Super18), in which the Bay of Plenty is (most unfortunately, depending on your disposition) drafted under the Waikato region's entry with The Chiefs. Being the most popular sport and widely announced news topic in the country means that even people that don't particuarly follow rugby know this, and simple trolls (whether intentional or not) cannot be allowed to slide past quality control; for those that do can get quite upset about this kind of thing. Plus international folk can't be getting the wrong idea; Everyone knows the internet is no place for misinformation!