Makes 'singletons' available on window.ss
We're assigning instances to the 'ss' namespace because singletons only work
within the context on a single Browserify bundle.

For example - assume the 'lib' bundle exposes a singleton called 'register'.
If bundle 'a' imports 'register', as an external dependency, then all modules
in bundle 'a' will get the same copy of 'register' when importing it.

Likewise if bundle 'b' imports 'register' as an external dependency,
all modules in bundle 'b' will get the same copy of 'register'.

However the copy of 'register' available to all modules in bundle 'a' is not
the same copy of 'register' that's available to modules in bundle 'b'.
Singletons only work within the context of a Browserify bundle but
not across bundles.