Don't call fetchRecord() in CampaignAdminList if already loaded
React can destroy and re-mount components by its own internal decisioning based on what's outdated,
which shouldn't trigger an XHR each time if the data is already present.

The CampaignAdminList mapStateToProps() call is getting the current record from the state.records list, which can be populated in two ways:
1. All campaigns are loaded for the overview (with all detail record data required for CampaignAdminList) - uses fetchRecords()
2. Only current campaign is loaded when URL is called directly (admin/campaigns/show/<id>) - uses fetchRecord()
As long as we ensure that fetchRecords() will always return the full data set for all records, this approach works fine.